correcting sed statement

+++ firstboot.py
@@ -21,7 +21,7 @@ if yesno == "y" or debugmode == False:
 
 # Enable PiScreenDrivers and configure calibration settings
 # calibration steps come from http://ozzmaker.com/forums/topic/piscreen-raspberripi2-touchscreen-calibration/
-yesno = raw_input("Enable Arduino uploading? (y/n) ")
+yesno = raw_input("Enable Piscreen? (y/n) ")
 if yesno == "y" or debugmode == False:
     #h.cmdcall("sudo sed -i """$ a\dtoverlay=piscreen,speed=16000000,rotate=90""" /boot/config.txt")
     sp.call(["sudo", "sed", "-i", "$ a\dtoverlay=piscreen,speed=16000000,rotate=90", "/boot/config.txt"])
@@ -34,8 +34,9 @@ if yesno == "y" or debugmode == False:
     #h.cmdcall("sed -i /^DISPLAY=*xinput*(?i)Touchscreen*(?i)Evdev Axis (?i)Inversion*$ /s/^/#/ /etc/X11/xinit/xinitrc")
     sp.call(["sudo", "sed", "-i", "/^DISPLAY=*xinput*(?i)Touchscreen*(?i)Evdev (?i)Axis (?i)Inversion*$ /s/^/#/", "/etc/X11/xinit/xinitrc"])
     # sp.call(shellcmd.split())
-    h.cmdcall("sudo mv etc/pointercal.xinput etc/pointercal.xinput_copy")
-    # sp.call(shellcmd.split())
+    if os.path.isfile("etc/pointercal.xinput"):
+        h.cmdcall("sudo mv etc/pointercal.xinput etc/pointercal.xinput_copy")
+        # sp.call(shellcmd.split())
 
 # upgrade packages
 yesno = raw_input("Upgrade packages? (y/n) ")
